Consolidate wildcard lookups in ScalaPositionManager
See:
  - https://github.com/JetBrains/intellij-scala/commit/b5faf8cffd2a812e476f2a891a799860b795a9ee
  - https://github.com/JetBrains/intellij-scala/pull/386

The code added to optionally use a full scan of
VM.allClasses when computing `getAllClasses`
when FORCE_POSITION_LOOKUP_IN_NESTED_TYPES is set overlaps
with existing functionality in `NamePattern.checkParts`.

The latter seems more efficient and more robust.
For example, `getNestedClasses` currently fails
to recognize `Outer$Inner` as a nested type of
`Outer$`.

This commit re-uses `NamePattern` as the implementation
for FORCE_POSITION_LOOKUP_IN_NESTED_TYPES.

I believe that the use case is for projects that have
compiler-plugin or macro-annotation driven code
synthesis where IntelliJ is unable to reason about
potential synthetic classes wrapping the user-written
code.

I have an external test case for this. It would be
good to integrate into the the plugin's test suite.

https://github.com/retronym/boxer/blob/topic/xasync/README.md